Slightly different than the original, we had refactored some of this code on trunk.  size'edness issues in sdbm, which is limited to int's (and in one case is manipulating an array of short[]'s.)
Backport: r587439

git-svn-id: https://svn.apache.org/repos/asf/apr/apr-util/branches/1.2.x@587617 13f79535-47bb-0310-9956-ffa450edef68

+++ dbm/sdbm/sdbm_pair.c
@@ -183,7 +183,7 @@ apr_sdbm_datum_t key;
 		register int m;
 		register char *dst = pag + (i == 1 ? PBLKSIZ : ino[i - 1]);
 		register char *src = pag + ino[i + 1];
-		register int   zoo = dst - src;
+		register short zoo = (short) (dst - src);
 
 		debug(("free-up %d ", zoo));
 /*
